Deprecate module standardizers
