+++ tests/conftest.py
@@ -45,12 +45,12 @@ def features_dataset() -> xr.Dataset:
 def features_multi() -> xr.Dataset:
     """
     Create a dataset as if it was loaded from `features.zarr`.
-    Coherent with the number of data transformations defined in the standardizers file.
+    Coherent with the number of data transformations defined in the normalizers module.
     """
-    import mlpp_lib.standardizers as st
+    import mlpp_lib.normalizers as no
 
     rng = np.random.default_rng(1)
-    X = rng.standard_normal(size=(*SHAPE, len([n.name for n in st.DataTransformation.__subclasses__()])))
+    X = rng.standard_normal(size=(*SHAPE, len([n.name for n in no.DataTransformation.__subclasses__()])))
     X = np.float64(X)
     X[(X > 4.5) | (X < -4.5)] = np.nan
     features = xr.Dataset(
@@ -73,9 +73,9 @@ def datatransformations() -> list:
     Create a list of data transformations.
     The list consists of all available data transformations.
     """
-    import mlpp_lib.standardizers as st
+    import mlpp_lib.normalizers as no
 
-    datatransformations = [st.create_transformation_from_str(n.name) for n in st.DataTransformation.__subclasses__()]
+    datatransformations = [no.create_transformation_from_str(n.name) for n in no.DataTransformation.__subclasses__()]
 
     return datatransformations
 
@@ -85,11 +85,11 @@ def data_transformer() -> xr.Dataset:
     """
     Create a datatransformer.
     """
-    import mlpp_lib.standardizers as st
+    import mlpp_lib.normalizers as no
 
-    transformations_list = [n.name for n in st.DataTransformation.__subclasses__()]
+    transformations_list = [n.name for n in no.DataTransformation.__subclasses__()]
     method_var_dict = {transformation: ([f"var{i}"],{}) for i, transformation in enumerate(transformations_list)}
-    data_transformer = st.DataTransformer(method_var_dict)
+    data_transformer = no.DataTransformer(method_var_dict)
 
     return data_transformer
